Move OsoError back to polar module
